// lifetrace-frontend/lib/generated/schemas/journalAutoLinkRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * LifeTrace API
 * OpenAPI spec version: 0.1.2
 */

/**
 * 自动关联请求
 */
export interface JournalAutoLinkRequest {
  /** 日记ID */
  journal_id?: number | null;
  /** 日记标题 */
  title?: string | null;
  /** 日记原文 */
  content_original?: string | null;
  /** 日记日期 */
  date: string;
  /** 日记归属刷新点 */
  day_bucket_start?: string | null;
  /**
   * 默认关联数量
   * @minimum 1
   * @maximum 10
   */
  max_items?: number;
}