// lifetrace-frontend/lib/generated/schemas/journalAutoLinkResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * LifeTrace API
 * OpenAPI spec version: 0.1.2
 */
import type { JournalAutoLinkCandidate } from './journalAutoLinkCandidate';

/**
 * 自动关联响应
 */
export interface JournalAutoLinkResponse {
  /** 关联待办ID列表 */
  related_todo_ids?: number[];
  /** 关联活动ID列表 */
  related_activity_ids?: number[];
  /** 待办候选 */
  todo_candidates?: JournalAutoLinkCandidate[];
  /** 活动候选 */
  activity_candidates?: JournalAutoLinkCandidate[];
}