// lifetrace-frontend/lib/generated/schemas/journalGenerateRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * LifeTrace API
 * OpenAPI spec version: 0.1.2
 */

/**
 * 生成客观记录/AI 视角请求
 */
export interface JournalGenerateRequest {
  /** 日记ID */
  journal_id?: number | null;
  /** 日记标题 */
  title?: string | null;
  /** 日记原文 */
  content_original?: string | null;
  /** 日记日期 */
  date?: string | null;
  /** 日记归属刷新点 */
  day_bucket_start?: string | null;
  /**
   * 语言
   * @maxLength 10
   */
  language?: string;
}