// lifetrace-frontend/lib/generated/schemas/journalListResponse.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LifeTrace API
 * LifeTrace API
 * OpenAPI spec version: 0.1.2
 */
import type { JournalResponse } from './journalResponse';

/**
 * 日记列表响应模型
 */
export interface JournalListResponse {
  /** 总数 */
  total: number;
  /** 日记列表 */
  journals: JournalResponse[];
}